Update to Go 1.18
