+++ go.mod
@@ -1,9 +1,10 @@
 module github.com/go-toolsmith/astcopy
 
-go 1.16
+go 1.18
 
 require (
-	github.com/go-toolsmith/astequal v1.1.0
+	github.com/go-toolsmith/astequal v1.2.0
 	github.com/go-toolsmith/strparse v1.1.0
-	golang.org/x/exp/typeparams v0.0.0-20230203172020-98cc5a0785f9
 )
+
+require github.com/google/go-cmp v0.6.0 // indirect
